pyproject.toml
[tool.poetry]
name = "youtube-statistics"
version = "0.1.0"
description = "YouTube Analytics with Telegram Bot interface"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "domain" },
    { include = "youtube" },
    { include = "report" },
    { include = "telegram_bot" }
]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "^2.25.2"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.111.0"
python-telegram-bot = "^20.7"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
ipython = "^8.18.1"

[tool.poetry.scripts]
youtube-stats = "main:run"
telegram-bot = "telegram_bot.bot:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true